data-scripts/requirements.txt
smartapi-python>=1.3.0
logzero>=1.7.0
NorenRestApiPy==0.0.22
pandas>=2.0.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
pyotp>=2.9.0
tqdm>=4.66.0
requests>=2.31.0
